임시..

+++ src/main/java/com/brogrammers/brogrammers/web/products/AllProductController.java
@@ -1,6 +1,7 @@
 package com.brogrammers.brogrammers.web.products;
 
 import com.brogrammers.brogrammers.domain.member.Member;
+import com.brogrammers.brogrammers.domain.order.OrderProducts;
 import com.brogrammers.brogrammers.domain.product.*;
 import com.brogrammers.brogrammers.domain.service.*;
 import com.brogrammers.brogrammers.form.DuplicatedProduct;
@@ -39,6 +40,7 @@ public class AllProductController {
     @Autowired CategoryService categoryService;
     @Autowired BrandService brandService;
     @Autowired ProductCategoryService productCategoryService;
+    @Autowired OrderProductsService orderProductsService;
     /////////////////////////////////////상품 등록 ///////////////////////////////////
     @GetMapping("/products/add")
     public String addForm(@ModelAttribute("form") ProductForm form ,Model model, @PathVariable(value = "productId",required = false)Long productId,
@@ -102,7 +104,7 @@ public class AllProductController {
 @GetMapping("/products/list")  // 전체 상품 조회
 public String productList(Model model, HttpServletRequest request,
                           @PageableDefault(page=0,size=12,sort="id",direction = Sort.Direction.DESC)Pageable pageable,
-                          String nameSearch,Long brandId,Long categoryId){
+                          String nameSearch,Long brandId,Long categoryId,String gender){
 
 
     /// 페이징 처리 /////////////////////////
@@ -111,30 +113,64 @@ public String productList(Model model, HttpServletRequest request,
         if(brandId!=null){ // 키워드O 브랜드O
             Brand brand = brandService.findById(brandId).get();
             if(categoryId!=null){// 키워드O 카테고리O 브랜드O
-                products = productCategoryService.findProductsByBrandNameCategory(brand,nameSearch,categoryService.findById(categoryId).get(),pageable);
+                if(!gender.isEmpty()){ // 키워드O 카테고리O 브랜드O 성별O
+                    products = productCategoryService.findProductsByBrandNameCategoryGender(brand,nameSearch,categoryService.findById(categoryId).get(),pageable,gender);
+                }else{ // 키워드O 카테고리O 브랜드O 성별X
+                    products = productCategoryService.findProductsByBrandNameCategory(brand,nameSearch,categoryService.findById(categoryId).get(),pageable);
+                }
             } else{ // 키워드O 브랜드O 카테고리X
-                products = productService.findProductsByBrandAndKeyword(nameSearch,brand,pageable);
+                if(!gender.isEmpty()){ // 키워드O 브랜드O 카테고리X 성별 O
+                    products = productService.findProductsByBrandAndKeywordAndGender(nameSearch,brand,gender,pageable);
+                } else{ // 키워드O 브랜드O 카테고리X 성별 X
+                    products = productService.findProductsByBrandAndKeyword(nameSearch,brand,pageable);
+                }
             }
         }else{ // 키워드 OOO 브랜드XXXX
             if(categoryId!=null){ // 키워드O 브랜드X 카테고리O
-                products = productCategoryService.findProductsByNameCategory(nameSearch,categoryService.findById(categoryId).get(),pageable);
+                if(!gender.isEmpty()){ // 키워드O 브랜드X 카테고리O 성별O
+                    products = productCategoryService.findProductsByNameCategoryGender(nameSearch,categoryService.findById(categoryId).get(),gender,pageable);
+                }else{ // 키워드O 브랜드X 카테고리O 성별X
+                    products = productCategoryService.findProductsByNameCategory(nameSearch,categoryService.findById(categoryId).get(),pageable);
+                }
             } else{ // 키워드O 브랜드X 카테고리X
-                products = productService.productSearchList(nameSearch,pageable);
+                if(!gender.isEmpty()){  // 키워드O 브랜드X 카테고리X 성별O
+                    products = productService.productSearchListGender(nameSearch,gender,pageable);
+                }else{ //  // 키워드O 브랜드X 카테고리X 성별X
+                    products = productService.productSearchList(nameSearch,pageable);
+                }
             }
         }
     }else{ // 키워드 XXXXXXXXXX
         if(brandId!=null){ // 브랜드 OOOOOOO
             Brand brand = brandService.findById(brandId).get();
             if(categoryId!=null){  // 키워드X 브랜드O 카테고리O
-                products = productCategoryService.findProductsByBrandCategory(brand,categoryService.findById(categoryId).get(),pageable);
-            }else{  // 키워드X 브랜드O 카테고리X
-                products = productService.findProductsByBrand(brand,pageable);
+                if(gender!=null){   // 키워드X 브랜드O 카테고리O 성별O
+                    products = productCategoryService.findProductsByBrandCategory(brand,categoryService.findById(categoryId).get(),pageable);
+                } else{ // 키워드X 브랜드O 카테고리O 성별X
+                    products = productCategoryService.findProductsByBrandCategoryGender(brand,categoryService.findById(categoryId).get(),gender,pageable);
+                }
+            }else{  // 키워드X 브랜드O 카테고리X 성별O
+                if(gender!=null){
+                    products = productService.findProductsByBrandGender(brand,gender,pageable);
+                } else{ // 키워드X 브랜드O 카테고리X 성별X
+                    products = productService.findProductsByBrand(brand,pageable);
+                }
             }
         }else{ // 검색 키워드가 없고 브랜드 조건이 없을 때
             if(categoryId!=null){ // 키워드x 브랜드x 카테고리O
-                products = productCategoryService.findProductsByCategory1(categoryService.findById(categoryId).get(),pageable);
+                if(gender!=null){ // 키워드x 브랜드x 카테고리O 성별O
+                    products = productCategoryService.findProductsByCategoryGender(categoryService.findById(categoryId).get(),gender,pageable);
+                } else{
+                    products = productCategoryService.findProductsByCategory1(categoryService.findById(categoryId).get(),pageable);
+                }
             } else{
-                products = productService.findAll(pageable);
+                if(gender!=null){ // 키워드X 브랜드X 카테고리X 성별O
+                    products = productService.findProductsByGender(gender,pageable);
+                } else{ // 키워드X 브랜드X 카테고리X 성별X
+
+                    products = productService.findAll(pageable);
+
+                }
             }
         }
     }
@@ -149,6 +185,7 @@ public String productList(Model model, HttpServletRequest request,
     String pname = "list";
     List<Category> categories = categoryService.findAll();
     List<Brand> brands = brandService.findAll();
+    model.addAttribute("gender",gender);
     model.addAttribute("totalPage",products.getTotalPages());
     model.addAttribute("bodytitle",bodytitle);
     model.addAttribute("pname",pname);
@@ -158,6 +195,9 @@ public String productList(Model model, HttpServletRequest request,
     model.addAttribute("endPage",endPage);
     model.addAttribute("categories", categories);
     model.addAttribute("brands", brands);
+
+    // 헤더 home, style, shop중 shop 글자 굵기를 위한
+    model.addAttribute("shop", "shop");
 
     return "products/list";
 }
@@ -284,13 +324,17 @@ public String productList(Model model, HttpServletRequest request,
     ///////////////////////////////////// edit ///////////////////////////////////
 
     @GetMapping("/products/delete/{id}")
-    public String delete(@PathVariable("id") Long id, HttpServletRequest request){
+    public String delete(@PathVariable("id") Long id, HttpServletRequest request,Model model){
         Member member = fun.getMember(request);
         if(member == null || member.getAccessrigths().equals("NORMAL")){
             return "redirect:/";
         }
-        System.out.println("상품의 아이디 : " + id);
         Products products = productService.getByid(id);
+        List<OrderProducts> orderProductsByProducts = orderProductsService.findOrderProductsByProducts(products);
+        if(orderProductsByProducts != null){
+            model.addAttribute("productId",id);
+            return "alert/errorDeleteProduct";
+        }
         List<Imgs> imgs = imgService.findImgsByProducts(products);
         if(products.getImgPath()!=null){
             new File(products.getImgPath()).delete();
